Cloud: scene store tabs, one frames page, single-row mobile search bar
- Frames home: the search box and the + button stay on one row at every
  width (the stacked @md layout left a tall empty band on mobile).
- Scene store: "Public scene store" / "My private scenes" tabs. The private
  list (filters, zip upload, visibility/delete) moved from /account/scenes
  (alias /scenes) to /my-scenes on the scenes host; the old URLs redirect.
- /frames is the only frames page: /account/frames redirects there (the
  workspace confirms pending enrollments and revokes itself). AccountNav
  drops the frames/scenes tabs; /account lands on Backends.
- Root → /frames: on the prod cloud host via the surface router; in the
  single-origin dev setup the store front answers at /store instead.
