+++ cloud/apps/auth-web/app/account/page.tsx
@@ -2,8 +2,9 @@ import { redirect } from "next/navigation";
 import { getAccountUrl } from "../../src/lib/env";
 
 // The overview page is gone — its headline numbers live in the account
-// header as tags. Land on the first real section instead: frames, which is
-// what the account is mostly about and the first tab in the nav.
+// header as tags. Land on the first real section instead. Frames and scenes
+// have their own surfaces (the /frames workspace and the scene store), so the
+// account is about backends, backups, activity and security.
 export default function AccountPage() {
-  redirect(getAccountUrl("/account/frames"));
+  redirect(getAccountUrl("/account/installs"));
 }
